+++ R/generics.R
@@ -1,0 +1,11 @@
+summary.compareR <- function(object, ...) {
+  UseMethod("summary")
+}
+
+print.compareR <- function(x, ...) {
+  UseMethod("print")
+}
+
+plot.compareR <- function(x, ...) {
+  UseMethod("plot")
+}